wip, making it classy
